requirements.txt
requests==2.28.1
pytest==7.1.2
allure-pytest==2.9.47
pyyaml==6.0
jsonpath-ng==1.7.0